move errors in one module
